refactor(core/engine/preflight): split run_preflight_in_place_with_options into SRP submodules
`run_preflight_in_place_with_options` was a 316-line orchestrator that
mixed seven distinct phases (event-sink setup, git cleanliness, registry
client init, workspace dry-run, per-package dry-run, per-package
registry/regime/ownership checks, finishability + duration estimation).

Mirror the existing `engine/publish/` submodule pattern (anticipated by
`engine/CLAUDE.md`) and split into:

- `engine/preflight/mod.rs` — orchestrator `run()` + `PreflightRunOptions`
- `engine/preflight/dry_run.rs` — workspace + per-package dry-run
- `engine/preflight/package_check.rs` — per-package registry/regime/ownership
- `engine/preflight/duration.rs` — `estimate_preflight_duration`

Public API is unchanged: `engine::run_preflight*` entry points become
one-line delegations; `engine::PreflightRunOptions` is re-exported from
the new submodule so existing callers (notably `shipper-cli`) keep working.

No behavior change. Event ordering, sidecar paths, and error messages
preserved verbatim. All 33 preflight tests pass; clippy clean.
